chore(test): reconcile duplicate test files exposed by relocation
Three test files collided with existing tests after the relocation
cleanup. Each pair was independently reviewed for coverage,
infrastructure quality, and overlap, then consolidated:

- editors.test.ts merged at src/client/test/components/calendar/
  (37 tests). Adopts module-level vi.mock pattern, AuthenticationService
  mock, DOM-driven assertions, and i18n-key style. Drops conditional-
  guard test pattern that silently passed on absent methods.

- moderation-settings.test.ts merged at src/client/test/components/admin/
  (16 tests, matching the broader admin tests convention). Adopts real
  i18next.init() over key-passthrough mock, controlled-promise loading
  pattern, and proper teardown.

- ActivityPub schemas.test.ts split into schemas-base.test.ts (35
  tests) and schemas-activities.test.ts (91 tests) under
  src/server/activitypub/test/validation/. Base file ports A's unique
  recipientSchema negatives, objectReferenceSchema, timestampSchema,
  and activityTypeSchema coverage on top of B's stronger bodies.
