+++ src/client/test/components/calendar/editors.test.ts
@@ -1,43 +1,96 @@
-import { expect, describe, it, beforeEach, afterEach, vi } from 'vitest';
-import { createMemoryHistory, createRouter, Router } from 'vue-router';
-import { RouteRecordRaw } from 'vue-router';
-import { flushPromises } from '@vue/test-utils';
+import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { mount, flushPromises } from '@vue/test-utils';
+import { nextTick } from 'vue';
 
 import { CalendarEditor } from '@/common/model/calendar_editor';
-import { mountComponent } from '@/client/test/lib/vue';
-import Editors from '@/client/components/logged_in/calendar-management/editors.vue';
+import {
+  CalendarEditorPermissionError,
+  EditorAlreadyExistsError,
+  EditorNotFoundError,
+} from '@/common/exceptions/editor';
+import { AccountInviteAlreadyExistsError } from '@/common/exceptions';
+
+import EditorsTab from '@/client/components/logged_in/calendar-management/editors.vue';
 import CalendarService from '@/client/service/calendar';
-import { CalendarEditorPermissionError, EditorAlreadyExistsError, EditorNotFoundError } from '@/common/exceptions/editor';
+import AuthenticationService from '@/client/service/authn';
 
-const routes: RouteRecordRaw[] = [
-  { path: '/test', component: {}, name: 'test' },
-];
+// Stub for PillButton component — preserves the variant class so tests can
+// distinguish primary vs danger buttons in the same way the real component does.
+const PillButtonStub = {
+  template: '<button :class="[\'pill-button\', `pill-button--${variant}`]" @click="$emit(\'click\', $event)"><slot /></button>',
+  props: ['variant', 'type', 'disabled', 'size'],
+  emits: ['click'],
+};
 
-const mountEditorsComponent = (calendarId: string = 'test-calendar', isOwner: boolean = true) => {
-  const router: Router = createRouter({
-    history: createMemoryHistory(),
-    routes: routes,
-  });
+// Stub for lucide-vue-next icons.
+const IconStub = { template: '<span />' };
 
-  const wrapper = mountComponent(Editors, router, {
-    props: {
-      calendarId: calendarId,
-      isOwner: isOwner,
+// Shared global stubs reused by every test mount in this suite. Hoisted to a
+// constant rather than repeated per-test to keep the suite scannable.
+const globalStubs = {
+  ModalLayout: {
+    template: '<div class="modal-stub"><slot /></div>',
+  },
+  EmptyLayout: {
+    template: '<div class="empty-stub"><slot /></div>',
+  },
+  AdminSectionHeader: {
+    template: '<div class="admin-section-header"><slot name="actions" /></div>',
+    props: ['title'],
+  },
+  LoadingMessage: {
+    template: '<div class="loading-message" />',
+    props: ['description'],
+  },
+  PillButton: PillButtonStub,
+  Plus: IconStub,
+  Mail: IconStub,
+  Crown: IconStub,
+  Globe: IconStub,
+  ArrowUp: IconStub,
+  Trash2: IconStub,
+  X: IconStub,
+};
+
+// Mock i18next-vue so translation keys come back as raw keys (with params
+// JSON-appended). Tests assert against keys instead of localized strings,
+// which keeps them stable when copy changes.
+vi.mock('i18next-vue', () => ({
+  useTranslation: () => ({
+    t: (key: string, params?: any) => {
+      if (params) {
+        return `${key}:${JSON.stringify(params)}`;
+      }
+      return key;
     },
-  });
+  }),
+}));
 
-  return {
-    wrapper,
-    router,
-  };
+// Mock services at module level so every `new CalendarService()` /
+// `new AuthenticationService()` in the component returns our stub instance.
+vi.mock('@/client/service/calendar');
+vi.mock('@/client/service/authn');
+
+const mountEditors = (props: Record<string, any> = {}, extraOptions: Record<string, any> = {}) => {
+  return mount(EditorsTab, {
+    props: {
+      calendarId: 'test-calendar',
+      ...props,
+    },
+    global: {
+      stubs: globalStubs,
+    },
+    ...extraOptions,
+  });
 };
 
 describe('Editors Component', () => {
   let calendarServiceMock: any;
-  let currentWrapper: any = null;
+  let authServiceMock: any;
 
   beforeEach(() => {
-    // Mock the calendar service
+    vi.clearAllMocks();
+
     calendarServiceMock = {
       listCalendarEditors: vi.fn(),
       grantEditAccess: vi.fn(),
@@ -45,30 +98,15 @@ describe('Editors Component', () => {
       cancelInvitation: vi.fn(),
       resendInvitation: vi.fn(),
     };
+    vi.mocked(CalendarService).mockImplementation(() => calendarServiceMock);
 
-    // Replace the CalendarService constructor with our mock
-    vi.spyOn(CalendarService.prototype, 'listCalendarEditors').mockImplementation(
-      calendarServiceMock.listCalendarEditors,
-    );
-    vi.spyOn(CalendarService.prototype, 'grantEditAccess').mockImplementation(
-      calendarServiceMock.grantEditAccess,
-    );
-    vi.spyOn(CalendarService.prototype, 'revokeEditAccess').mockImplementation(
-      calendarServiceMock.revokeEditAccess,
-    );
-    vi.spyOn(CalendarService.prototype, 'cancelInvitation').mockImplementation(
-      calendarServiceMock.cancelInvitation,
-    );
-    vi.spyOn(CalendarService.prototype, 'resendInvitation').mockImplementation(
-      calendarServiceMock.resendInvitation,
-    );
+    authServiceMock = {
+      userEmail: vi.fn().mockReturnValue(null),
+    };
+    vi.mocked(AuthenticationService).mockImplementation(() => authServiceMock);
   });
 
   afterEach(() => {
-    if (currentWrapper) {
-      currentWrapper.unmount();
-      currentWrapper = null;
-    }
     vi.restoreAllMocks();
   });
 
@@ -88,10 +126,8 @@ describe('Editors Component', () => {
         pendingInvitations: testPendingInvitations,
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await flushPromises();
 
       expect(calendarServiceMock.listCalendarEditors).toHaveBeenCalledWith('test-calendar');
       expect(wrapper.vm.state.editors).toEqual(testEditors);
@@ -100,20 +136,15 @@ describe('Editors Component', () => {
     });
 
     it('displays loading state while fetching editors', async () => {
-      // Mock with a promise that never resolves to keep loading state
+      // Mock with a promise that never resolves to keep loading state.
       const neverResolvingPromise = new Promise(() => {});
       calendarServiceMock.listCalendarEditors.mockReturnValue(neverResolvingPromise);
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for the onMounted hook to execute
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await nextTick();
 
       expect(wrapper.vm.state.isLoading).toBe(true);
       expect(wrapper.find('.loading-message').exists()).toBe(true);
-
-      // Note: We don't resolve this promise since we want to test the loading state
     });
 
     it('handles permission error when loading editors', async () => {
@@ -121,13 +152,10 @@ describe('Editors Component', () => {
         new CalendarEditorPermissionError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for async error handling to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
-      expect(wrapper.vm.state.error).toBe('You don\'t have permission to manage editors for this calendar');
+      expect(wrapper.vm.state.error).toBe('error_permission_denied');
       expect(wrapper.find('.alert--error').exists()).toBe(true);
     });
 
@@ -136,26 +164,22 @@ describe('Editors Component', () => {
         new CalendarEditorPermissionError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors();
       await flushPromises();
 
-      // The permission-denied error message should be visible
+      // The permission-denied error message should be visible.
       expect(wrapper.find('.alert--error').exists()).toBe(true);
-      // The "No editors" empty state must NOT be shown alongside the error
-      expect(wrapper.find('.empty-state').exists()).toBe(false);
+      // The empty state must NOT be shown alongside the error.
+      expect(wrapper.find('.empty-stub').exists()).toBe(false);
     });
 
     it('handles generic error when loading editors', async () => {
       calendarServiceMock.listCalendarEditors.mockRejectedValue(new Error('Network error'));
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
 
-      await wrapper.vm.$nextTick();
-
-      expect(wrapper.vm.state.error).toBe('Failed to load calendar editors');
+      expect(wrapper.vm.state.error).toBe('error_loading_editors');
     });
   });
 
@@ -171,10 +195,7 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for async loading to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
       const editorItems = wrapper.findAll('.editor-card');
@@ -189,14 +210,10 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for async loading to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
-      expect(wrapper.find('.empty-state').exists()).toBe(true);
-      expect(wrapper.find('.empty-state').text()).toContain('No editors have been granted access to this calendar');
+      expect(wrapper.find('.empty-stub').exists()).toBe(true);
     });
 
     it('shows remove button for each editor', async () => {
@@ -209,15 +226,45 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for async loading to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
       const removeButton = wrapper.find('.editor-actions .btn-ghost');
       expect(removeButton.exists()).toBe(true);
-      expect(removeButton.text()).toBe('Remove');
+      expect(removeButton.text()).toBe('remove_button');
+    });
+
+    it('should display pending invitations separately from active editors', async () => {
+      const mockEditors = [
+        CalendarEditor.fromObject({
+          id: 'editor-1',
+          calendarId: 'calendar-123',
+          email: 'alice@example.com',
+        }),
+      ];
+
+      const mockInvitations = [
+        {
+          id: 'invite-1',
+          email: 'charlie@example.com',
+        },
+      ];
+
+      calendarServiceMock.listCalendarEditors.mockResolvedValue({
+        activeEditors: mockEditors,
+        pendingInvitations: mockInvitations,
+      });
+
+      const wrapper = mountEditors({ calendarId: 'calendar-123' });
+      await flushPromises();
+      await nextTick();
+
+      // Should have sections for both active editors and pending invitations.
+      const sections = wrapper.findAll('.editors-section');
+      expect(sections.length).toBeGreaterThanOrEqual(2);
+
+      expect(wrapper.html()).toContain('alice@example.com');
+      expect(wrapper.html()).toContain('charlie@example.com');
     });
   });
 
@@ -230,23 +277,19 @@ describe('Editors Component', () => {
     });
 
     it('shows Add Editor button in empty state when isOwner is true', async () => {
-      const { wrapper } = mountEditorsComponent('test-calendar', true);
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors({ isOwner: true });
       await flushPromises();
 
-      // Empty state PillButton should be present for owners
+      // Empty state PillButton should be present for owners.
       const pillButtons = wrapper.findAll('.pill-button--primary');
       expect(pillButtons.length).toBeGreaterThan(0);
     });
 
     it('hides Add Editor button in empty state when isOwner is false', async () => {
-      const { wrapper } = mountEditorsComponent('test-calendar', false);
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors({ isOwner: false });
       await flushPromises();
 
-      // No primary pill button should appear for non-owners in empty state
+      // No primary pill button should appear for non-owners in empty state.
       const pillButtons = wrapper.findAll('.pill-button--primary');
       expect(pillButtons.length).toBe(0);
     });
@@ -257,9 +300,7 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent('test-calendar', true);
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors({ isOwner: true });
       await flushPromises();
 
       const headerAddButton = wrapper.find('.admin-section-header .pill-button--primary');
@@ -272,9 +313,7 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent('test-calendar', false);
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors({ isOwner: false });
       await flushPromises();
 
       const headerAddButton = wrapper.find('.admin-section-header .pill-button--primary');
@@ -282,29 +321,26 @@ describe('Editors Component', () => {
     });
 
     it('defaults isOwner to false when prop is not provided', async () => {
-      const router: Router = createRouter({
-        history: createMemoryHistory(),
-        routes: routes,
-      });
-
-      const wrapper = mountComponent(Editors, router, {
+      // Mount without an explicit isOwner prop — the component default should
+      // suppress the Add Editor primary button.
+      const wrapper = mount(EditorsTab, {
         props: {
           calendarId: 'test-calendar',
-          // isOwner not provided — should default to false
+        },
+        global: {
+          stubs: globalStubs,
         },
       });
-      currentWrapper = wrapper;
 
       await flushPromises();
 
-      // No primary pill button should appear by default
       const pillButtons = wrapper.findAll('.pill-button--primary');
       expect(pillButtons.length).toBe(0);
     });
   });
 
   describe('Add Editor Functionality', () => {
-    beforeEach(async () => {
+    beforeEach(() => {
       calendarServiceMock.listCalendarEditors.mockResolvedValue({
         activeEditors: [],
         pendingInvitations: [],
@@ -312,71 +348,88 @@ describe('Editors Component', () => {
     });
 
     it('opens add editor form when button clicked', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors({ isOwner: true });
       await flushPromises();
+      await nextTick();
 
       const addButton = wrapper.find('.pill-button--primary');
+      expect(addButton.exists()).toBe(true);
+
       await addButton.trigger('click');
+      await nextTick();
 
       expect(wrapper.vm.state.showAddForm).toBe(true);
-      // Check that the modal wrapper exists
-      expect(wrapper.find('[data-testid="modal-layout"]').exists() ||
-             wrapper.findComponent({ name: 'ModalLayout' }).exists() ||
-             wrapper.find('.add-editor-form').exists()).toBe(true);
+      expect(wrapper.find('.add-editor-form').exists()).toBe(true);
+
+      // Email input should be rendered inside the form.
+      const input = wrapper.find('#email');
+      expect(input.exists()).toBe(true);
+    });
+
+    it('should focus email input when modal opens', async () => {
+      const wrapper = mount(EditorsTab, {
+        props: {
+          calendarId: 'test-calendar',
+          isOwner: true,
+        },
+        attachTo: document.body,
+        global: {
+          stubs: globalStubs,
+        },
+      });
+
+      await flushPromises();
+      await nextTick();
+
+      await wrapper.find('.pill-button--primary').trigger('click');
+      await nextTick();
+
+      const input = wrapper.find('#email').element as HTMLInputElement;
+      expect(document.activeElement).toBe(input);
+
+      wrapper.unmount();
     });
 
     it('closes add editor form when cancel clicked', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
       // Open the form
       wrapper.vm.state.showAddForm = true;
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
-      // Call the close method directly since button text may be translated
+      // Call the close method directly since button text is a translation key.
       wrapper.vm.closeAddForm();
 
       expect(wrapper.vm.state.showAddForm).toBe(false);
     });
 
     it('adds editor successfully', async () => {
-      const testEditor = new CalendarEditor('1', 'test-calendar', 'existing@example.com');
       const newEditor = new CalendarEditor('3', 'test-calendar', 'newuser@example.com');
       calendarServiceMock.grantEditAccess.mockResolvedValue(newEditor);
 
-      // Mock the subsequent loadEditors call to return the updated list
-      calendarServiceMock.listCalendarEditors.mockResolvedValueOnce({
-        activeEditors: [testEditor],
-        pendingInvitations: [],
-      }).mockResolvedValueOnce({
-        activeEditors: [testEditor, newEditor],
-        pendingInvitations: [],
-      });
+      // First call resolves with empty list (mount), second resolves with the
+      // newly added editor (post-grant reload).
+      calendarServiceMock.listCalendarEditors
+        .mockResolvedValueOnce({ activeEditors: [], pendingInvitations: [] })
+        .mockResolvedValueOnce({ activeEditors: [newEditor], pendingInvitations: [] });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors({ isOwner: true });
       await flushPromises();
+      await nextTick();
 
-      // Open add form first
-      wrapper.vm.state.showAddForm = true;
-      wrapper.vm.state.newAccountId = 'newuser@example.com';
-      await wrapper.vm.$nextTick();
+      // Open add form via the visible primary pill button.
+      await wrapper.find('.pill-button--primary').trigger('click');
+      await nextTick();
 
-      // Verify form is open before adding
-      expect(wrapper.vm.state.showAddForm).toBe(true);
+      // Fill in the email field via real DOM interaction.
+      const input = wrapper.find('#email');
+      await input.setValue('newuser@example.com');
+      await nextTick();
 
-      // Call add method directly
-      await wrapper.vm.addEditor();
-
-      // Wait for add operation to complete
+      // Submit by clicking the form's primary button.
+      const submitButton = wrapper.find('.add-editor-form .pill-button--primary');
+      await submitButton.trigger('click');
       await flushPromises();
 
       expect(calendarServiceMock.grantEditAccess).toHaveBeenCalledWith(
@@ -388,19 +441,17 @@ describe('Editors Component', () => {
     });
 
     it('handles empty account ID error', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
 
-      await wrapper.vm.$nextTick();
-
-      // Open add form with empty account ID
+      // Open add form with empty account ID.
       wrapper.vm.state.showAddForm = true;
       wrapper.vm.state.newAccountId = '';
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
       await wrapper.vm.addEditor();
 
-      expect(wrapper.vm.state.addError).toBe('Account ID is required');
+      expect(wrapper.vm.state.addError).toBe('account_id_required');
       expect(calendarServiceMock.grantEditAccess).not.toHaveBeenCalled();
     });
 
@@ -409,18 +460,46 @@ describe('Editors Component', () => {
         new EditorAlreadyExistsError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors({ isOwner: true });
+      await flushPromises();
+      await nextTick();
 
-      await wrapper.vm.$nextTick();
+      // Open add form via the visible primary pill button.
+      await wrapper.find('.pill-button--primary').trigger('click');
+      await nextTick();
 
-      wrapper.vm.state.showAddForm = true;
-      wrapper.vm.state.newAccountId = 'existing@example.com';
-      await wrapper.vm.$nextTick();
+      await wrapper.find('#email').setValue('existing@example.com');
+      await nextTick();
 
-      await wrapper.vm.addEditor();
+      const submitButton = wrapper.find('.add-editor-form .pill-button--primary');
+      await submitButton.trigger('click');
+      await flushPromises();
 
-      expect(wrapper.vm.state.addError).toBe('This user is already an editor on this calendar.');
+      expect(wrapper.vm.state.addError).toBe('error_editor_already_exists');
+      expect(wrapper.vm.state.showAddForm).toBe(true); // Form should stay open.
+    });
+
+    it('handles invitation already exists error', async () => {
+      calendarServiceMock.grantEditAccess.mockRejectedValue(
+        new AccountInviteAlreadyExistsError(),
+      );
+
+      const wrapper = mountEditors({ isOwner: true });
+      await flushPromises();
+      await nextTick();
+
+      await wrapper.find('.pill-button--primary').trigger('click');
+      await nextTick();
+
+      await wrapper.find('#email').setValue('invited@example.com');
+      await nextTick();
+
+      const submitButton = wrapper.find('.add-editor-form .pill-button--primary');
+      await submitButton.trigger('click');
+      await flushPromises();
+
+      expect(wrapper.vm.state.addError).toBe('error_invite_already_exists');
+      expect(wrapper.vm.state.showAddForm).toBe(true);
     });
 
     it('handles permission error when adding editor', async () => {
@@ -428,26 +507,27 @@ describe('Editors Component', () => {
         new CalendarEditorPermissionError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await flushPromises();
 
       wrapper.vm.state.showAddForm = true;
       wrapper.vm.state.newAccountId = 'user@example.com';
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
       await wrapper.vm.addEditor();
 
-      expect(wrapper.vm.state.addError).toBe('You don\'t have permission to manage editors for this calendar');
+      expect(wrapper.vm.state.addError).toBe('error_permission_denied');
     });
   });
 
   describe('Remove Editor Functionality', () => {
     let testEditor: CalendarEditor;
 
-    beforeEach(async () => {
+    beforeEach(() => {
       testEditor = new CalendarEditor('1', 'test-calendar', 'user1@example.com');
+      // Set the current user to someone OTHER than the editor under test so
+      // the leave-calendar UI doesn't render and confuse the .btn-ghost lookup.
+      authServiceMock.userEmail.mockReturnValue('owner@example.com');
       calendarServiceMock.listCalendarEditors.mockResolvedValue({
         activeEditors: [testEditor],
         pendingInvitations: [],
@@ -455,21 +535,25 @@ describe('Editors Component', () => {
     });
 
     it('shows confirmation modal when remove clicked', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
+      await nextTick();
 
-      await wrapper.vm.$nextTick();
+      // Click the visible remove button on the editor card.
+      const removeButton = wrapper.find('.editor-actions .btn-ghost');
+      expect(removeButton.exists()).toBe(true);
 
-      await wrapper.vm.confirmRemoveEditor(testEditor);
+      await removeButton.trigger('click');
+      await nextTick();
 
-      expect(wrapper.vm.state.editorToRemove).toEqual(testEditor);
+      expect(wrapper.vm.state.editorToRemove).not.toBeNull();
+      expect(wrapper.vm.state.editorToRemove.id).toBe('1');
+      expect(wrapper.find('.confirmation-modal').exists()).toBe(true);
     });
 
     it('cancels remove operation', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await flushPromises();
 
       wrapper.vm.state.editorToRemove = testEditor;
       await wrapper.vm.cancelRemoveEditor();
@@ -479,27 +563,32 @@ describe('Editors Component', () => {
 
     it('removes editor successfully', async () => {
       calendarServiceMock.revokeEditAccess.mockResolvedValue(undefined);
-      // Mock the subsequent loadEditors call to return empty list after removal
-      calendarServiceMock.listCalendarEditors.mockResolvedValueOnce({
-        activeEditors: [testEditor],
-        pendingInvitations: [],
-      }).mockResolvedValueOnce({
-        activeEditors: [],
-        pendingInvitations: [],
-      });
+      // First listCalendarEditors call (mount) returns the editor; reload after
+      // removal returns the empty list.
+      calendarServiceMock.listCalendarEditors
+        .mockResolvedValueOnce({ activeEditors: [testEditor], pendingInvitations: [] })
+        .mockResolvedValueOnce({ activeEditors: [], pendingInvitations: [] });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
+      await nextTick();
 
-      await wrapper.vm.$nextTick();
+      // Open the confirmation modal by clicking remove.
+      await wrapper.find('.editor-actions .btn-ghost').trigger('click');
+      await nextTick();
 
-      wrapper.vm.state.editorToRemove = testEditor;
+      // Confirm by invoking removeEditor (the danger button is inside the
+      // ModalLayout stub, so calling the method is the cleanest path).
       await wrapper.vm.removeEditor();
+      await flushPromises();
 
       expect(calendarServiceMock.revokeEditAccess).toHaveBeenCalledWith(
         'test-calendar',
         '1',
       );
+      // listCalendarEditors should be called twice: once on mount, once after
+      // the successful revoke.
+      expect(calendarServiceMock.listCalendarEditors).toHaveBeenCalledTimes(2);
       expect(wrapper.vm.state.editors).toHaveLength(0);
       expect(wrapper.vm.state.editorToRemove).toBeNull();
     });
@@ -509,16 +598,15 @@ describe('Editors Component', () => {
         new EditorNotFoundError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await flushPromises();
 
       wrapper.vm.state.editorToRemove = testEditor;
       await wrapper.vm.removeEditor();
 
-      expect(wrapper.vm.state.error).toBe('Editor not found');
-      expect(wrapper.vm.state.editors).toHaveLength(0); // Should remove from local list
+      expect(wrapper.vm.state.error).toBe('error_editor_not_found');
+      // Should remove from local list to keep UI in sync with server.
+      expect(wrapper.vm.state.editors).toHaveLength(0);
       expect(wrapper.vm.state.editorToRemove).toBeNull();
     });
 
@@ -527,20 +615,74 @@ describe('Editors Component', () => {
         new CalendarEditorPermissionError(),
       );
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      await wrapper.vm.$nextTick();
+      const wrapper = mountEditors();
+      await flushPromises();
 
       wrapper.vm.state.editorToRemove = testEditor;
       await wrapper.vm.removeEditor();
 
-      expect(wrapper.vm.state.error).toBe('You don\'t have permission to manage editors for this calendar');
+      expect(wrapper.vm.state.error).toBe('error_permission_denied');
+    });
+  });
+
+  describe('Editor Role Display', () => {
+    it('should distinguish current user as editor vs owner', async () => {
+      const mockEditors = [
+        CalendarEditor.fromObject({
+          id: 'editor-1',
+          calendarId: 'calendar-123',
+          email: 'alice@example.com',
+        }),
+      ];
+
+      // Set current user email to match an editor — they can leave the calendar.
+      authServiceMock.userEmail.mockReturnValue('alice@example.com');
+
+      calendarServiceMock.listCalendarEditors.mockResolvedValue({
+        activeEditors: mockEditors,
+        pendingInvitations: [],
+      });
+
+      const wrapper = mountEditors({ calendarId: 'calendar-123' });
+      await flushPromises();
+      await nextTick();
+
+      // Current user should be able to leave (is an editor).
+      expect(wrapper.vm.canLeaveCalendar()).toBe(true);
+
+      // Leave button should be visible.
+      const leaveButton = wrapper.find('.btn-ghost--danger');
+      expect(leaveButton.exists()).toBe(true);
+    });
+
+    it('should not show leave button if user is not an editor', async () => {
+      const mockEditors = [
+        CalendarEditor.fromObject({
+          id: 'editor-1',
+          calendarId: 'calendar-123',
+          email: 'bob@example.com',
+        }),
+      ];
+
+      // Current user is the owner (not in editors list).
+      authServiceMock.userEmail.mockReturnValue('owner@example.com');
+
+      calendarServiceMock.listCalendarEditors.mockResolvedValue({
+        activeEditors: mockEditors,
+        pendingInvitations: [],
+      });
+
+      const wrapper = mountEditors({ calendarId: 'calendar-123' });
+      await flushPromises();
+      await nextTick();
+
+      expect(wrapper.vm.canLeaveCalendar()).toBe(false);
+      expect(wrapper.find('.btn-ghost--danger').exists()).toBe(false);
     });
   });
 
   describe('UI State Management', () => {
-    beforeEach(async () => {
+    beforeEach(() => {
       calendarServiceMock.listCalendarEditors.mockResolvedValue({
         activeEditors: [],
         pendingInvitations: [],
@@ -548,18 +690,15 @@ describe('Editors Component', () => {
     });
 
     it('disables add button while loading', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
-      // Set loading state which should disable the button in empty state
+      // Set loading state which should disable the button in empty state.
       wrapper.vm.state.isLoading = true;
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
       // During loading, the empty state PillButton is not rendered (v-else),
-      // so verify loading state is active
+      // so verify loading state is active.
       expect(wrapper.vm.state.isLoading).toBe(true);
     });
 
@@ -570,33 +709,28 @@ describe('Editors Component', () => {
         pendingInvitations: [],
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
       wrapper.vm.state.isRemoving = testEditor.id;
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
       const removeButton = wrapper.find('.editor-actions .btn-ghost');
       expect(removeButton.attributes('disabled')).toBeDefined();
     });
 
     it('prevents form closure during operations', async () => {
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
-      // Wait for initial load to complete
+      const wrapper = mountEditors();
       await flushPromises();
 
       wrapper.vm.state.showAddForm = true;
       wrapper.vm.state.isAdding = true;
-      await wrapper.vm.$nextTick();
+      await nextTick();
 
       await wrapper.vm.closeAddForm();
 
-      expect(wrapper.vm.state.showAddForm).toBe(true); // Should not close
+      // Should not close while an add is in progress.
+      expect(wrapper.vm.state.showAddForm).toBe(true);
     });
   });
 
@@ -618,40 +752,34 @@ describe('Editors Component', () => {
         pendingInvitations: [testInvitation],
       });
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors();
       await flushPromises();
 
       expect(wrapper.vm.state.pendingInvitations).toEqual([testInvitation]);
-      // Note: UI display will be tested after component implementation
     });
 
     it('cancels invitation successfully', async () => {
-      calendarServiceMock.listCalendarEditors.mockResolvedValue({
-        activeEditors: [],
-        pendingInvitations: [testInvitation],
-      });
+      calendarServiceMock.listCalendarEditors
+        .mockResolvedValueOnce({ activeEditors: [], pendingInvitations: [testInvitation] })
+        .mockResolvedValueOnce({ activeEditors: [], pendingInvitations: [] });
       calendarServiceMock.cancelInvitation.mockResolvedValue(undefined);
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
+      await nextTick();
 
+      // The cancel button is the non-primary .btn-text inside the invitation card.
+      const cancelButton = wrapper.find('.btn-text:not(.btn-text--primary)');
+      expect(cancelButton.exists()).toBe(true);
+
+      await cancelButton.trigger('click');
       await flushPromises();
 
-      // Mock the reload to show invitation removed
-      calendarServiceMock.listCalendarEditors.mockResolvedValueOnce({
-        activeEditors: [],
-        pendingInvitations: [],
-      });
-
-      // Call cancel method (will be implemented in component)
-      if (wrapper.vm.cancelInvitation) {
-        await wrapper.vm.cancelInvitation('inv-1');
-
-        expect(calendarServiceMock.cancelInvitation).toHaveBeenCalledWith('test-calendar', 'inv-1');
-        expect(wrapper.vm.state.pendingInvitations).toEqual([]);
-      }
+      expect(calendarServiceMock.cancelInvitation).toHaveBeenCalledWith(
+        'test-calendar',
+        'inv-1',
+      );
+      expect(calendarServiceMock.listCalendarEditors).toHaveBeenCalledTimes(2);
     });
 
     it('resends invitation successfully', async () => {
@@ -661,17 +789,22 @@ describe('Editors Component', () => {
       });
       calendarServiceMock.resendInvitation.mockResolvedValue(undefined);
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
+      const wrapper = mountEditors();
+      await flushPromises();
+      await nextTick();
 
+      const resendButton = wrapper.find('.btn-text--primary');
+      expect(resendButton.exists()).toBe(true);
+
+      await resendButton.trigger('click');
       await flushPromises();
 
-      // Call resend method (will be implemented in component)
-      if (wrapper.vm.resendInvitation) {
-        await wrapper.vm.resendInvitation('inv-1');
-
-        expect(calendarServiceMock.resendInvitation).toHaveBeenCalledWith('test-calendar', 'inv-1');
-      }
+      expect(calendarServiceMock.resendInvitation).toHaveBeenCalledWith(
+        'test-calendar',
+        'inv-1',
+      );
+      // Component sets state.success on resend success.
+      expect(wrapper.vm.state.success).toBeTruthy();
     });
 
     it('handles cancel invitation errors', async () => {
@@ -681,17 +814,13 @@ describe('Editors Component', () => {
       });
       calendarServiceMock.cancelInvitation.mockRejectedValue(new Error('Network error'));
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors();
       await flushPromises();
 
-      // Call cancel method and expect error handling (will be implemented in component)
-      if (wrapper.vm.cancelInvitation) {
-        await wrapper.vm.cancelInvitation('inv-1');
+      await wrapper.vm.cancelInvitation('inv-1');
+      await flushPromises();
 
-        expect(wrapper.vm.state.error).toBeTruthy();
-      }
+      expect(wrapper.vm.state.error).toBeTruthy();
     });
 
     it('handles resend invitation errors', async () => {
@@ -701,17 +830,13 @@ describe('Editors Component', () => {
       });
       calendarServiceMock.resendInvitation.mockRejectedValue(new Error('Rate limited'));
 
-      const { wrapper } = mountEditorsComponent();
-      currentWrapper = wrapper;
-
+      const wrapper = mountEditors();
       await flushPromises();
 
-      // Call resend method and expect error handling (will be implemented in component)
-      if (wrapper.vm.resendInvitation) {
-        await wrapper.vm.resendInvitation('inv-1');
+      await wrapper.vm.resendInvitation('inv-1');
+      await flushPromises();
 
-        expect(wrapper.vm.state.error).toBeTruthy();
-      }
+      expect(wrapper.vm.state.error).toBeTruthy();
     });
   });
 });
