+++ src/server/activitypub/test/validation/schemas-activities.test.ts
@@ -1,270 +1,25 @@
+/**
+ * Tests for ActivityPub validation schemas - activity-type schemas.
+ *
+ * Covers the seven activity-type schemas (Follow, Accept, Create, Update,
+ * Delete, Announce, Undo) that extend activityBaseSchema. The base/shared
+ * schemas they build on are exercised in the sibling schemas-base.test.ts
+ * file.
+ */
+
 import { describe, it, expect } from 'vitest';
 
 import {
-  actorUriSchema,
-  objectUriSchema,
-  activityBaseSchema,
-  contextSchema,
-  recipientSchema,
+  followActivitySchema,
+  acceptActivitySchema,
   createActivitySchema,
   updateActivitySchema,
   deleteActivitySchema,
-  followActivitySchema,
-  acceptActivitySchema,
   announceActivitySchema,
   undoActivitySchema,
 } from '@/server/activitypub/validation/schemas';
 
-describe('ActivityPub Validation Schemas', () => {
-  describe('actorUriSchema', () => {
-    it('should accept valid HTTPS URLs', () => {
-      const validUrls = [
-        'https://example.com/users/alice',
-        'https://mastodon.social/@user',
-        'https://remote.instance/calendars/test-calendar',
-        'https://sub.domain.example.com/path/to/actor',
-      ];
-
-      for (const url of validUrls) {
-        const result = actorUriSchema.safeParse(url);
-        expect(result.success).toBe(true);
-        if (result.success) {
-          expect(result.data).toBe(url);
-        }
-      }
-    });
-
-    it('should allow HTTP URLs in test environment', () => {
-      // HTTP URLs are allowed in test/development environments
-      const result = actorUriSchema.safeParse('http://example.com/users/alice');
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.data).toBe('http://example.com/users/alice');
-      }
-    });
-
-    it('should reject invalid URLs', () => {
-      const invalidUrls = [
-        'not-a-url',
-        '',
-        'example.com/users/alice',
-      ];
-
-      for (const url of invalidUrls) {
-        const result = actorUriSchema.safeParse(url);
-        expect(result.success, `Expected ${url} to be rejected`).toBe(false);
-      }
-    });
-
-    it('should accept FTP URLs in test environment (too permissive)', () => {
-      // Note: In test/dev environments, the schema accepts any valid URL scheme
-      // This is probably too permissive but reflects current schema behavior
-      const result = actorUriSchema.safeParse('ftp://example.com/user');
-      expect(result.success).toBe(true);
-    });
-
-    it('should reject non-string values', () => {
-      const invalidValues = [123, null, undefined, {}, []];
-
-      for (const value of invalidValues) {
-        const result = actorUriSchema.safeParse(value);
-        expect(result.success).toBe(false);
-      }
-    });
-  });
-
-  describe('objectUriSchema', () => {
-    it('should accept valid URLs', () => {
-      const validUrls = [
-        'https://example.com/objects/123',
-        'https://remote.instance/events/abc-def',
-        'http://localhost:3000/test', // HTTP allowed for development
-      ];
-
-      for (const url of validUrls) {
-        const result = objectUriSchema.safeParse(url);
-        expect(result.success).toBe(true);
-        if (result.success) {
-          expect(result.data).toBe(url);
-        }
-      }
-    });
-
-    it('should reject invalid URLs', () => {
-      const invalidUrls = [
-        'not-a-url',
-        '',
-        'example.com/objects/123',
-      ];
-
-      for (const url of invalidUrls) {
-        const result = objectUriSchema.safeParse(url);
-        expect(result.success).toBe(false);
-      }
-    });
-  });
-
-  describe('contextSchema', () => {
-    it('should accept a single string', () => {
-      const result = contextSchema.safeParse('https://www.w3.org/ns/activitystreams');
-      expect(result.success).toBe(true);
-    });
-
-    it('should accept an array of strings', () => {
-      const result = contextSchema.safeParse([
-        'https://www.w3.org/ns/activitystreams',
-        'https://w3id.org/security/v1',
-      ]);
-      expect(result.success).toBe(true);
-    });
-
-    it('should accept an array with objects', () => {
-      const result = contextSchema.safeParse([
-        'https://www.w3.org/ns/activitystreams',
-        { '@language': 'en' },
-      ]);
-      expect(result.success).toBe(true);
-    });
-
-    it('should reject non-string, non-array values', () => {
-      const result = contextSchema.safeParse(123);
-      expect(result.success).toBe(false);
-    });
-  });
-
-  describe('recipientSchema', () => {
-    it('should accept a single string', () => {
-      const result = recipientSchema.safeParse('https://www.w3.org/ns/activitystreams#Public');
-      expect(result.success).toBe(true);
-    });
-
-    it('should accept an array of strings', () => {
-      const result = recipientSchema.safeParse([
-        'https://example.com/users/alice',
-        'https://example.com/users/bob',
-      ]);
-      expect(result.success).toBe(true);
-    });
-
-    it('should accept undefined (optional)', () => {
-      const result = recipientSchema.safeParse(undefined);
-      expect(result.success).toBe(true);
-    });
-  });
-
-  describe('activityBaseSchema', () => {
-    it('should accept a valid activity', () => {
-      const validActivity = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        id: 'https://example.com/activities/123',
-        type: 'Create',
-        actor: 'https://example.com/users/alice',
-        to: ['https://www.w3.org/ns/activitystreams#Public'],
-        cc: ['https://example.com/users/alice/followers'],
-      };
-
-      const result = activityBaseSchema.safeParse(validActivity);
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.data.type).toBe('Create');
-        expect(result.data.actor).toBe('https://example.com/users/alice');
-      }
-    });
-
-    it('should accept an activity with array context', () => {
-      const validActivity = {
-        '@context': [
-          'https://www.w3.org/ns/activitystreams',
-          'https://w3id.org/security/v1',
-        ],
-        id: 'https://example.com/activities/123',
-        type: 'Follow',
-        actor: 'https://example.com/users/alice',
-      };
-
-      const result = activityBaseSchema.safeParse(validActivity);
-      expect(result.success).toBe(true);
-    });
-
-    it('should accept an activity without to/cc fields', () => {
-      const validActivity = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        id: 'https://example.com/activities/123',
-        type: 'Accept',
-        actor: 'https://example.com/users/alice',
-      };
-
-      const result = activityBaseSchema.safeParse(validActivity);
-      expect(result.success).toBe(true);
-    });
-
-    it('should allow HTTP actor in test environment', () => {
-      const validActivity = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        id: 'https://example.com/activities/123',
-        type: 'Create',
-        actor: 'http://example.com/users/alice', // HTTP allowed in test environment
-      };
-
-      const result = activityBaseSchema.safeParse(validActivity);
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.data.actor).toBe('http://example.com/users/alice');
-      }
-    });
-
-    it('should reject an activity without required fields', () => {
-      const missingFields = [
-        { '@context': 'https://www.w3.org/ns/activitystreams', id: 'https://example.com/1', type: 'Create' }, // missing actor
-        { '@context': 'https://www.w3.org/ns/activitystreams', id: 'https://example.com/1', actor: 'https://example.com/users/alice' }, // missing type
-        { '@context': 'https://www.w3.org/ns/activitystreams', type: 'Create', actor: 'https://example.com/users/alice' }, // missing id
-        { id: 'https://example.com/1', type: 'Create', actor: 'https://example.com/users/alice' }, // missing @context
-      ];
-
-      for (const activity of missingFields) {
-        const result = activityBaseSchema.safeParse(activity);
-        expect(result.success).toBe(false);
-      }
-    });
-
-    it('should reject an activity with empty type', () => {
-      const invalidActivity = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        id: 'https://example.com/activities/123',
-        type: '',
-        actor: 'https://example.com/users/alice',
-      };
-
-      const result = activityBaseSchema.safeParse(invalidActivity);
-      expect(result.success).toBe(false);
-      if (!result.success) {
-        const typeError = result.error.issues.find(
-          (issue) => issue.path.includes('type'),
-        );
-        expect(typeError).toBeDefined();
-        expect(typeError?.message).toBe('Activity type is required');
-      }
-    });
-
-    it('should reject an activity with invalid id URL', () => {
-      const invalidActivity = {
-        '@context': 'https://www.w3.org/ns/activitystreams',
-        id: 'not-a-valid-url',
-        type: 'Create',
-        actor: 'https://example.com/users/alice',
-      };
-
-      const result = activityBaseSchema.safeParse(invalidActivity);
-      expect(result.success).toBe(false);
-      if (!result.success) {
-        const idError = result.error.issues.find(
-          (issue) => issue.path.includes('id'),
-        );
-        expect(idError).toBeDefined();
-      }
-    });
-  });
-
+describe('ActivityPub Validation Schemas - Activities', () => {
   describe('followActivitySchema', () => {
     it('should accept a valid Follow activity', () => {
       const validFollow = {
